+++ package/bin/bin.hash
@@ -1,0 +1,1 @@
+sha256  d2f96418893ac66156d0e691cda189b0d85ae1d814065d1d9aa1845383100f46  LICENSE
