+++ src/bin/autogen.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+autoreconf -W portability -vifm
